+++ common/src/main/java/com/telepathicgrunt/the_bumblezone/events/entity/EntityHurtEvent.java
@@ -1,10 +1,10 @@
 package com.telepathicgrunt.the_bumblezone.events.entity;
 
-import com.telepathicgrunt.the_bumblezone.events.base.CancellableEventHandler;
+import com.telepathicgrunt.the_bumblezone.events.base.EventHandler;
 import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.entity.LivingEntity;
 
 public record EntityHurtEvent(LivingEntity entity, DamageSource source, float amount) {
 
-    public static final CancellableEventHandler<EntityHurtEvent> EVENT_LOWEST = new CancellableEventHandler<>();
+    public static final EventHandler<EntityHurtEvent> EVENT_LOWEST = new EventHandler<>();
 }