+++ neoforge/src/main/java/com/telepathicgrunt/the_bumblezone/mixin/neoforge/item/ItemExtensionMixin.java
@@ -2,7 +2,7 @@ package com.telepathicgrunt.the_bumblezone.mixin.neoforge.item;
 
 import com.telepathicgrunt.the_bumblezone.platform.ItemExtension;
 import net.minecraft.world.item.ItemStack;
-import net.neoforged.neoforge.common.ToolAction;
+import net.neoforged.neoforge.common.ItemAbility;
 import net.neoforged.neoforge.common.extensions.IItemExtension;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
@@ -14,7 +14,7 @@ public interface ItemExtensionMixin extends IItemExtension {
     boolean bz$canPerformAction(ItemStack stack, String toolAction);
 
     @Override
-    default boolean canPerformAction(ItemStack stack, ToolAction toolAction) {
-        return this.bz$canPerformAction(stack, toolAction.name());
+    default boolean canPerformAction(ItemStack stack, ItemAbility itemAbility) {
+        return this.bz$canPerformAction(stack, itemAbility.name());
     }
 }