+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ##### Fluids:
 Improve the interaction between Honey Fluids and Lava.
 
+##### Items:
+(NeoForge): Made Crystal Cannon now tell Neo API that its default ammo is Honey Crystal Shards.
+
 
 ### **(V.7.6.4 Changes) (1.21 Minecraft)**